split each pipeline into 3 sub-pipelines
